refactor(gui): move menu into app module
